+++ sdk/dotnet/AVS/V20200717Preview/WorkloadNetworkDnsService.cs
@@ -61,7 +61,7 @@ namespace Pulumi.AzureNextGen.AVS.V20200717Preview
         /// NSX revision number.
         /// </summary>
         [Output("revision")]
-        public Output<int?> Revision { get; private set; } = null!;
+        public Output<double?> Revision { get; private set; } = null!;
 
         /// <summary>
         /// DNS Service status.
@@ -178,7 +178,7 @@ namespace Pulumi.AzureNextGen.AVS.V20200717Preview
         /// NSX revision number.
         /// </summary>
         [Input("revision")]
-        public Input<int>? Revision { get; set; }
+        public Input<double>? Revision { get; set; }
 
         public WorkloadNetworkDnsServiceArgs()
         {
