+++ sdk/dotnet/AVS/V20200717Preview/WorkloadNetworkSegment.cs
@@ -49,7 +49,7 @@ namespace Pulumi.AzureNextGen.AVS.V20200717Preview
         /// NSX revision number.
         /// </summary>
         [Output("revision")]
-        public Output<int?> Revision { get; private set; } = null!;
+        public Output<double?> Revision { get; private set; } = null!;
 
         /// <summary>
         /// Segment status.
@@ -142,7 +142,7 @@ namespace Pulumi.AzureNextGen.AVS.V20200717Preview
         /// NSX revision number.
         /// </summary>
         [Input("revision")]
-        public Input<int>? Revision { get; set; }
+        public Input<double>? Revision { get; set; }
 
         /// <summary>
         /// NSX Segment identifier. Generally the same as the Segment's display name
