+++ sdk/dotnet/Automation/Latest/Python2Package.cs
@@ -92,7 +92,7 @@ namespace Pulumi.AzureNextGen.Automation.Latest
         /// Gets or sets the size in bytes of the module.
         /// </summary>
         [Output("sizeInBytes")]
-        public Output<int?> SizeInBytes { get; private set; } = null!;
+        public Output<double?> SizeInBytes { get; private set; } = null!;
 
         /// <summary>
         /// Resource tags.
