+++ sdk/dotnet/Compute/V20180601/Disk.cs
@@ -25,7 +25,7 @@ namespace Pulumi.AzureNextGen.Compute.V20180601
         /// The number of IOPS allowed for this disk; only settable for UltraSSD disks. One operation can transfer between 4k and 256k bytes. For a description of the range of values you can set, see [Ultra SSD Managed Disk Offerings](https://docs.microsoft.com/azure/virtual-machines/windows/disks-ultra-ssd#ultra-ssd-managed-disk-offerings).
         /// </summary>
         [Output("diskIOPSReadWrite")]
-        public Output<int?> DiskIOPSReadWrite { get; private set; } = null!;
+        public Output<double?> DiskIOPSReadWrite { get; private set; } = null!;
 
         /// <summary>
         /// The bandwidth allowed for this disk; only settable for UltraSSD disks. MBps means millions of bytes per second - MB here uses the ISO notation, of powers of 10. For a description of the range of values you can set, see [Ultra SSD Managed Disk Offerings](https://docs.microsoft.com/azure/virtual-machines/windows/disks-ultra-ssd#ultra-ssd-managed-disk-offerings).
@@ -173,7 +173,7 @@ namespace Pulumi.AzureNextGen.Compute.V20180601
         /// The number of IOPS allowed for this disk; only settable for UltraSSD disks. One operation can transfer between 4k and 256k bytes. For a description of the range of values you can set, see [Ultra SSD Managed Disk Offerings](https://docs.microsoft.com/azure/virtual-machines/windows/disks-ultra-ssd#ultra-ssd-managed-disk-offerings).
         /// </summary>
         [Input("diskIOPSReadWrite")]
-        public Input<int>? DiskIOPSReadWrite { get; set; }
+        public Input<double>? DiskIOPSReadWrite { get; set; }
 
         /// <summary>
         /// The bandwidth allowed for this disk; only settable for UltraSSD disks. MBps means millions of bytes per second - MB here uses the ISO notation, of powers of 10. For a description of the range of values you can set, see [Ultra SSD Managed Disk Offerings](https://docs.microsoft.com/azure/virtual-machines/windows/disks-ultra-ssd#ultra-ssd-managed-disk-offerings).
