+++ sdk/dotnet/Compute/V20200630/Snapshot.cs
@@ -31,7 +31,7 @@ namespace Pulumi.AzureNextGen.Compute.V20200630
         /// The size of the disk in bytes. This field is read only.
         /// </summary>
         [Output("diskSizeBytes")]
-        public Output<int> DiskSizeBytes { get; private set; } = null!;
+        public Output<double> DiskSizeBytes { get; private set; } = null!;
 
         /// <summary>
         /// If creationData.createOption is Empty, this field is mandatory and it indicates the size of the disk to create. If this field is present for updates or creation with other options, it indicates a resize. Resizes are only allowed if the disk is not attached to a running VM, and can only increase the disk's size.
