+++ sdk/dotnet/Compute/V20201001Preview/Outputs/CloudServiceRoleSkuResponse.cs
@@ -16,7 +16,7 @@ namespace Pulumi.AzureNextGen.Compute.V20201001Preview.Outputs
         /// <summary>
         /// Specifies the number of role instances in the cloud service.
         /// </summary>
-        public readonly int? Capacity;
+        public readonly double? Capacity;
         /// <summary>
         /// The sku name. NOTE: If the new SKU is not supported on the hardware the cloud service is currently on, you need to delete and recreate the cloud service or move back to the old sku.
         /// </summary>
@@ -28,7 +28,7 @@ namespace Pulumi.AzureNextGen.Compute.V20201001Preview.Outputs
 
         [OutputConstructor]
         private CloudServiceRoleSkuResponse(
-            int? capacity,
+            double? capacity,
 
             string? name,
