+++ sdk/dotnet/DataBoxEdge/V20200501Preview/Device.cs
@@ -49,7 +49,7 @@ namespace Pulumi.AzureNextGen.DataBoxEdge.V20200501Preview
         /// The Data Box Edge/Gateway device local capacity in MB.
         /// </summary>
         [Output("deviceLocalCapacity")]
-        public Output<int> DeviceLocalCapacity { get; private set; } = null!;
+        public Output<double> DeviceLocalCapacity { get; private set; } = null!;
 
         /// <summary>
         /// The Data Box Edge/Gateway device model.
